requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flask-talisman==1.1.0
python-dotenv==1.2.2
requests==2.33.1
Werkzeug==3.1.8
gunicorn==25.3.0
feedparser==6.0.11
beautifulsoup4==4.13.4
